Added cbot_guidance package
