+++ cbot_guidance/launch/cbot_guidance.launch
@@ -1,0 +1,5 @@
+<?xml version="1.0"?>
+<launch>
+    <node name="guidance_node" pkg="cbot_guidance" type="guidance_node" output="screen">
+    </node>
+</launch>
